# CappingDataScience/api/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is documentation for the globetrotters DATA 477 capping group's\
    \ API."
  version: "1.0.0"
  title: "NBA API"
  contact:
    email: "nicholas.blaskey1@marist.edu"
  license:
    name: "MIT"
    url: "https://mit-license.org/"
host: "127.0.0.1"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /model/clustering:
    post:
      tags:
      - "model"
      summary: "Cluster a set of players"
      description: "Returns cluster IDs for the provided players"
      operationId: "model_clustering_post"
      parameters:
      - in: "body"
        name: "body"
        description: "Player ids to cluster"
        required: true
        schema:
          $ref: "#/definitions/ClusterInput"
      responses:
        "200":
          description: "Sucess"
          schema:
            $ref: "#/definitions/ClusterOutput"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.model_controller"
  /model/reinforcement:
    post:
      tags:
      - "model"
      summary: "Run a betting bot starting at a date"
      description: "Runs our betting bot based on input"
      operationId: "model_reinforcement_post"
      parameters:
      - in: "body"
        name: "body"
        description: "Input to configure the betting bot"
        required: true
        schema:
          $ref: "#/definitions/BettingBotInput"
      responses:
        "200":
          description: "Sucess"
          schema:
            $ref: "#/definitions/BettingBotOutput"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.model_controller"
  /model/neuralNetwork:
    post:
      tags:
      - "model"
      summary: "Runs a prediction on a given game using a possible supplied tweet."
      description: "Gives the chance of team winning from the first game at the date\
        \ supplied. Also provides features the nerual network is using to predict."
      operationId: "model_neural_network_post"
      parameters:
      - in: "body"
        name: "body"
        description: "Input to the neural network"
        required: true
        schema:
          $ref: "#/definitions/NerualNetInput"
      responses:
        "200":
          description: "Sucess"
          schema:
            $ref: "#/definitions/NerualNetOutput"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.model_controller"
  /model/ensemble:
    post:
      tags:
      - "model"
      summary: "Runs a prediction on a number of games using an ensemble approach."
      description: "Utilizes an ensemble approach using both the reinforcement and\
        \ the neural network."
      operationId: "model_ensemble_post"
      parameters:
      - in: "body"
        name: "body"
        description: "Input for the ensemble model."
        required: true
        schema:
          $ref: "#/definitions/EnsembleInput"
      responses:
        "200":
          description: "Sucess"
          schema:
            $ref: "#/definitions/EnsembleOutput"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.model_controller"
...
definitions:
...
  Player:
    type: "object"
    properties:
      player_auto_key:
        type: "integer"
        format: "int64"
      player_id:
        type: "integer"
        format: "int64"
      player_name:
        type: "string"
      height:
        type: "integer"
        format: "int64"
      weight:
        type: "integer"
        format: "int64"
      season:
        type: "integer"
        format: "int64"
      college:
        type: "string"
      age:
        type: "integer"
        format: "int64"
      country:
        type: "string"
      draft_year:
        type: "integer"
        format: "int64"
      draft_round:
        type: "integer"
        format: "int64"
      draft_number:
        type: "integer"
        format: "int64"
      team_id:
        type: "integer"
        format: "int64"
      season_range:
        type: "integer"
        format: "int64"
    example:
      college: "college"
      country: "country"
      draft_number: 3
      draft_year: 7
      weight: 5
      draft_round: 9
      player_auto_key: 0
      team_id: 2
      player_id: 6
      season: 5
      player_name: "player_name"
      season_range: 4
      age: 2
      height: 1
...
  Game:
    type: "object"
    properties:
      game_id:
        type: "integer"
        format: "int64"
      game_date:
        type: "string"
        format: "date-time"
      home_team_id:
        type: "integer"
        format: "int64"
      away_team_id:
        type: "integer"
        format: "int64"
      season:
        type: "integer"
        format: "int64"
      home_team_won:
        type: "boolean"
      home_first:
        type: "integer"
        format: "int64"
      home_second:
        type: "integer"
        format: "int64"
      home_third:
        type: "integer"
        format: "int64"
      home_fourth:
        type: "integer"
        format: "int64"
      home_open:
        type: "integer"
        format: "int64"
      home_close:
        type: "integer"
        format: "int64"
      home_ml:
        type: "integer"
        format: "int64"
      home_two_h:
        type: "integer"
        format: "int64"
      away_first:
        type: "integer"
        format: "int64"
      away_second:
        type: "integer"
        format: "int64"
      away_third:
        type: "integer"
        format: "int64"
      away_fourth:
        type: "integer"
        format: "int64"
      away_open:
        type: "integer"
        format: "int64"
      away_close:
        type: "integer"
        format: "int64"
      away_ml:
        type: "integer"
        format: "int64"
      away_two_h:
        type: "integer"
        format: "int64"
      home_won_last:
        type: "boolean"
      away_won_last:
        type: "boolean"
    example:
      home_third: 7
      home_close: 2
      away_third: 1
      home_team_id: 6
      away_open: 7
      home_team_won: true
      away_ml: 4
      away_first: 1
      away_close: 1
      away_won_last: true
      home_second: 2
      home_first: 5
      away_fourth: 6
      home_open: 3
      home_fourth: 9
      away_second: 1
      season: 5
      away_two_h: 5
      home_two_h: 7
      away_team_id: 1
      home_won_last: true
      game_date: "2000-01-23T04:56:07.000+00:00"
      home_ml: 4
      game_id: 0
...
  BettingBotInput:
    type: "object"
    properties:
      starting_money:
        type: "integer"
        format: "int64"
      starting_date:
        type: "string"
        format: "date-time"
      num_games:
        type: "integer"
        format: "int64"
  BettingBotOutput:
    type: "object"
    properties:
      games:
        type: "array"
        items:
          $ref: "#/definitions/Game"
      teams:
        type: "array"
        items:
          $ref: "#/definitions/Team"
      bet_on_games:
        type: "array"
        items:
          type: "boolean"
      ending_money:
        type: "integer"
        format: "int64"
    example:
      bet_on_games:
      - true
      - true
      teams:
      - owner: "owner"
        head_coach: "head_coach"
        manager: "manager"
        city: "city"
        arena_name: "arena_name"
        year_founded: 5
        team_id: 0
        abbreviation: "abbreviation"
        max_year_in_nba_champ: 1
        nickname: "nickname"
        min_year_in_nba_champ: 6
        arena_capacity: 5
        players_been_on_team: 2
      - owner: "owner"
        head_coach: "head_coach"
        manager: "manager"
        city: "city"
        arena_name: "arena_name"
        year_founded: 5
        team_id: 0
        abbreviation: "abbreviation"
        max_year_in_nba_champ: 1
        nickname: "nickname"
        min_year_in_nba_champ: 6
        arena_capacity: 5
        players_been_on_team: 2
      games:
      - home_third: 7
        home_close: 2
        away_third: 1
        home_team_id: 6
        away_open: 7
        home_team_won: true
        away_ml: 4
        away_first: 1
        away_close: 1
        away_won_last: true
        home_second: 2
        home_first: 5
        away_fourth: 6
        home_open: 3
        home_fourth: 9
        away_second: 1
        season: 5
        away_two_h: 5
        home_two_h: 7
        away_team_id: 1
        home_won_last: true
        game_date: "2000-01-23T04:56:07.000+00:00"
        home_ml: 4
        game_id: 0
      - home_third: 7
        home_close: 2
        away_third: 1
        home_team_id: 6
        away_open: 7
        home_team_won: true
        away_ml: 4
        away_first: 1
        away_close: 1
        away_won_last: true
        home_second: 2
        home_first: 5
        away_fourth: 6
        home_open: 3
        home_fourth: 9
        away_second: 1
        season: 5
        away_two_h: 5
        home_two_h: 7
        away_team_id: 1
        home_won_last: true
        game_date: "2000-01-23T04:56:07.000+00:00"
        home_ml: 4
        game_id: 0
      ending_money: 0
  ClusterInput:
    type: "object"
    properties:
      season:
        type: "integer"
        format: "int64"
      player_ids:
        type: "array"
        items:
          type: "integer"
  ClusterOutput:
    type: "object"
    properties:
      players_clustered:
        type: "array"
        items:
          $ref: "#/definitions/ClusterOutputItem"
      players:
        type: "array"
        items:
          $ref: "#/definitions/Player"
      missing_or_not_found:
        type: "array"
        items:
          type: "integer"
    example:
      missing_or_not_found:
      - 9
      - 9
      players:
      - college: "college"
        country: "country"
        draft_number: 3
        draft_year: 7
        weight: 5
        draft_round: 9
        player_auto_key: 0
        team_id: 2
        player_id: 6
        season: 5
        player_name: "player_name"
        season_range: 4
        age: 2
        height: 1
      - college: "college"
        country: "country"
        draft_number: 3
        draft_year: 7
        weight: 5
        draft_round: 9
        player_auto_key: 0
        team_id: 2
        player_id: 6
        season: 5
        player_name: "player_name"
        season_range: 4
        age: 2
        height: 1
      players_clustered:
      - player_id: 0
        average_assists: 1.46581298050294517310021547018550336360931396484375
        average_steals: 5.962133916683182377482808078639209270477294921875
        average_blocks: 5.63737665663332876420099637471139430999755859375
        defense_label: 7
        average_points: 6.02745618307040320615897144307382404804229736328125
        offense_label: 2
      - player_id: 0
        average_assists: 1.46581298050294517310021547018550336360931396484375
        average_steals: 5.962133916683182377482808078639209270477294921875
        average_blocks: 5.63737665663332876420099637471139430999755859375
        defense_label: 7
        average_points: 6.02745618307040320615897144307382404804229736328125
        offense_label: 2
  ClusterOutputItem:
    type: "object"
    properties:
      player_id:
        type: "integer"
        format: "int64"
      average_points:
        type: "number"
        format: "double"
      average_assists:
        type: "number"
        format: "double"
      average_steals:
        type: "number"
        format: "double"
      average_blocks:
        type: "number"
        format: "double"
      offense_label:
        type: "integer"
        format: "int64"
      defense_label:
        type: "integer"
        format: "int64"
    example:
      player_id: 0
      average_assists: 1.46581298050294517310021547018550336360931396484375
      average_steals: 5.962133916683182377482808078639209270477294921875
      average_blocks: 5.63737665663332876420099637471139430999755859375
      defense_label: 7
      average_points: 6.02745618307040320615897144307382404804229736328125
      offense_label: 2
...
  NerualNetInput:
    type: "object"
    properties:
      date:
        type: "string"
        format: "date-time"
      tweet:
        type: "string"
  NerualNetOutput:
    type: "object"
    properties:
      game:
        $ref: "#/definitions/Game"
      features:
        $ref: "#/definitions/NeuralNetFeatures"
      chance_to_win:
        type: "number"
        format: "double"
      actual_team_won:
        type: "boolean"
    example:
      features:
        prev_game_points: 5
        year: 6
        prev_game_three_throw_perc: 7.061401241503109105224211816675961017608642578125
        team: 1
        prev_game_rebounds: 3
        did_tweet: 7
        prev_game_free_throw_perc: 2.3021358869347654518833223846741020679473876953125
        prev_game_assists: 9
        is_home_team: 2
        month: 0
        prev_game_field_goal_perc: 5.63737665663332876420099637471139430999755859375
        tweet_sentiment: 4
        won_last_game: 1
      game:
        home_third: 7
        home_close: 2
        away_third: 1
        home_team_id: 6
        away_open: 7
        home_team_won: true
        away_ml: 4
        away_first: 1
        away_close: 1
        away_won_last: true
        home_second: 2
        home_first: 5
        away_fourth: 6
        home_open: 3
        home_fourth: 9
        away_second: 1
        season: 5
        away_two_h: 5
        home_two_h: 7
        away_team_id: 1
        home_won_last: true
        game_date: "2000-01-23T04:56:07.000+00:00"
        home_ml: 4
        game_id: 0
      actual_team_won: true
      chance_to_win: 1.024645700144157789424070870154537260532379150390625
...
  EnsembleInput:
    type: "object"
    properties:
      starting_money:
        type: "integer"
        format: "int64"
      starting_date:
        type: "string"
        format: "date-time"
      num_games:
        type: "integer"
        format: "int64"
      tweets:
        type: "array"
        items:
          type: "string"
  EnsembleOutput:
    type: "object"
    properties:
      games:
        type: "array"
        items:
          $ref: "#/definitions/Game"
      teams:
        type: "array"
        items:
          $ref: "#/definitions/Team"
      betting_bot_bet_on_games:
        type: "array"
        items:
          type: "boolean"
      neural_network_prob:
        type: "array"
        items:
          type: "number"
          format: "double"
    example:
      neural_network_prob:
      - 0.80082819046101150206595775671303272247314453125
      - 0.80082819046101150206595775671303272247314453125
      teams:
      - owner: "owner"
        head_coach: "head_coach"
        manager: "manager"
        city: "city"
        arena_name: "arena_name"
        year_founded: 5
        team_id: 0
        abbreviation: "abbreviation"
        max_year_in_nba_champ: 1
        nickname: "nickname"
        min_year_in_nba_champ: 6
        arena_capacity: 5
        players_been_on_team: 2
      - owner: "owner"
        head_coach: "head_coach"
        manager: "manager"
        city: "city"
        arena_name: "arena_name"
        year_founded: 5
        team_id: 0
        abbreviation: "abbreviation"
        max_year_in_nba_champ: 1
        nickname: "nickname"
        min_year_in_nba_champ: 6
        arena_capacity: 5
        players_been_on_team: 2
      games:
      - home_third: 7
        home_close: 2
        away_third: 1
        home_team_id: 6
        away_open: 7
        home_team_won: true
        away_ml: 4
        away_first: 1
        away_close: 1
        away_won_last: true
        home_second: 2
        home_first: 5
        away_fourth: 6
        home_open: 3
        home_fourth: 9
        away_second: 1
        season: 5
        away_two_h: 5
        home_two_h: 7
        away_team_id: 1
        home_won_last: true
        game_date: "2000-01-23T04:56:07.000+00:00"
        home_ml: 4
        game_id: 0
      - home_third: 7
        home_close: 2
        away_third: 1
        home_team_id: 6
        away_open: 7
        home_team_won: true
        away_ml: 4
        away_first: 1
        away_close: 1
        away_won_last: true
        home_second: 2
        home_first: 5
        away_fourth: 6
        home_open: 3
        home_fourth: 9
        away_second: 1
        season: 5
        away_two_h: 5
        home_two_h: 7
        away_team_id: 1
        home_won_last: true
        game_date: "2000-01-23T04:56:07.000+00:00"
        home_ml: 4
        game_id: 0
      betting_bot_bet_on_games:
      - true
      - true
...